Started work on issue #8 This code should work, but isn't tested
Update issue 8
Labels: Started
WifiScanner could be used to start an asynchronus scan. It is not possible to start a synchronus scan, because this is not supported by the Android Wifi API. 
The results are submitted to an instance of the interface WifiResultCallback. If the scan was successfull scanFinished is called. If an Exception occured, scanFailed with the Exception is called.

git-svn-id: https://wificompass.googlecode.com/svn/trunk@54 6f3e661d-c991-e48d-a97c-fec7ba324fcc