did some work with the PopupDrawable and added last known location to the Project Site
Update issue #16
DB version: 15
added Location to the ProjectSite to save the last known location of the user in this project

Update issue #9
refactored PopupDrawable code, so the Popup itself can handle, if it should be shown or not. Helps make our life easier, because we won't have to implement isActive in every superDrawable. This is something which belongs to the Popup, so we implement it there.

changed the MeasuringDrawable icon to be a little bigger, to be easier touchable. The icon itself wasn't scaled, only the transparent area arround

git-svn-id: https://wificompass.googlecode.com/svn/trunk@65 6f3e661d-c991-e48d-a97c-fec7ba324fcc