Update issue #11 when starting step detection, wifi scans are also started every 3 seconds. If we receive a result before 3 seconds, we just save the result and don't start a new one, because we don't want to have to frequent results. If we don't receive a result in time, we try to remind the driver, we wanted to get a result. Maybe it helps.
On a Samsung Galaxy Tab 1000 with only a few (1-2) WiFi networks around, the driver does not return any results even after minutes. This seams to be device and driver related. We can't affect this, but we send the asynchronous start scan to the wifiservice. 

If you are in a location with more wifi networks, the scanning with the Samsung GT1000 works fine though. 

git-svn-id: https://wificompass.googlecode.com/svn/trunk@130 6f3e661d-c991-e48d-a97c-fec7ba324fcc